% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_boxplot}
\alias{draw_boxplot}
\title{Draw a boxplot}
\usage{
draw_boxplot(x)
}
\arguments{
\item{x}{input dataframe with the required columns}
}
\value{
a boxplot from ggplot
}
\description{
Draw a boxplot of a dataframe with columns called \code{type} and \code{cortisol}
}
\examples{
draw_boxplot(births)



}
