% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_raw}
\alias{read_raw}
\title{Read raw data}
\usage{
read_raw(my_url = "http://ritsokiguess.site/datafiles/births_wide.csv")
}
\arguments{
\item{my_url}{Input url of wide data (optional)}
}
\value{
a tidy (longer) version of the wide data with missings removed
}
\description{
Read raw (wider) data from (by default) a URL, reformat it longer, remove
missings and translate numbered groups to birth types
}
\examples{
read_raw()

}
