% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_welch}
\alias{run_welch}
\title{Run Welch anova}
\usage{
run_welch(x)
}
\arguments{
\item{x}{input dataframe with required columns}
}
\value{
the output from Welch ANOVA plus Games-Howell test (in a \code{list})
}
\description{
Run Welch ANOVA (not assuming equal variances) on a dataframe with columns called \code{type} and \code{cortisol}
}
