% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thing2-package.R
\docType{package}
\name{thing2-package}
\alias{thing2}
\alias{thing2-package}
\title{thing2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ken Butler \email{ken.butler@utoronto.ca} (\href{https://orcid.org/0000-0001-7821-6811}{ORCID})

}
\keyword{internal}
